Move validateAuth into its own file
